Todo WEB APP

+++ app.js
@@ -1,17 +1,18 @@
 const express = require("express");
 const bodyParser = require("body-parser")
 const mongoose = require('mongoose')
+const _ = require('lodash')
 const app = express();
 var items = [];
 const PORT = 5000;
-app.set("view engine","ejs")
+app.set("view engine", "ejs")
 app.use(express.static("public"));
 app.use(bodyParser.urlencoded({ extended: true }));
-mongoose.connect("mongodb://localhost:27017/todolistDB");
+mongoose.connect("mongodb+srv://gausu-admin:Gausu123@cluster0.eykn3.mongodb.net/todolistDB");
 const itemsSchema = {
     name: String
 };
-const Item = new mongoose.model("Item",itemsSchema);
+const Item = new mongoose.model("Item", itemsSchema);
 const item1 = new Item({
     name: "Welcome to ToDo List App."
 });
@@ -21,7 +22,13 @@ const item2 = new Item({
 const item3 = new Item({
     name: "Happy Coding!"
 });
-const defaulItems = [item1,item2,item3];
+const defaulItems = [item1, item2, item3];
+const listSchema = {
+    name: String,
+    items: [itemsSchema]
+};
+const List = new mongoose.model("list",listSchema);
+
 // Item.insertMany(defaulItems,(err)=>{
 //     if (err) {
 //         console.log(err);
@@ -39,29 +46,85 @@ app.get("/", (req, res) => {
         day: "numeric",
         month: "long"
     };
-    var day = today.toLocaleDateString("en-US",option);
-    Item.find({},(err,items)=>{
-        res.render("list",{kindOfDay:day,newListItem:items})
+    var day = today.toLocaleDateString("en-US", option);
+    Item.find({}, (err, items) => {
+        res.render("list", { kindOfDay: day, newListItem: items })
     })
-    
+
 });
-app.post("/",(req,res)=>{
+app.post("/", (req, res) => {
+    var today = new Date();
+    var option = {
+        weekday: "long",
+        day: "numeric",
+        month: "long"
+    };
+    var day = today.toLocaleDateString("en-US", option);
     const itemName = req.body.newItem;
+    const listName = req.body.list;
     const item = new Item({
         name: itemName,
     });
+    if (listName === day) {
+        item.save();
+        res.redirect("/");
+        
+    } else {
+        List.findOne({name:listName},(err,foundList)=>{
+            foundList.items.push(item);
+            foundList.save();
+            res.redirect("/"+listName);
+        })
+    }
 
-    item.save();
-    res.redirect("/");
+    
 })
-app.post("/delete",(req,res)=>{
-    const checkedItemId = req.body.checkbox;
-    Item.findByIdAndRemove(checkedItemId,(err)=>{
+app.get("/:customListName",(req,res)=>{
+    const customListName = _.capitalize(req.params.customListName);
+    List.findOne({name:customListName},(err,foundList)=>{
         if (!err) {
-            console.log("Successfully Deleted Item!")
-            res.redirect('/')   
-        }
+            if (!foundList) {
+                const list = new List({
+                    name: customListName,
+                    items: defaulItems
+                });
+                list.save();
+                
+            } else {
+                res.render("list",{ kindOfDay: foundList.name, newListItem: foundList.items });
+                
+            }  
+        } 
     })
+    
+})
+app.post("/delete", (req, res) => {
+    var today = new Date();
+    var option = {
+        weekday: "long",
+        day: "numeric",
+        month: "long"
+    };
+    var day = today.toLocaleDateString("en-US", option);
+    const checkedItemId = req.body.checkbox;
+    const listName = req.body.listName;
+    if (listName === day) {
+        Item.findByIdAndRemove(checkedItemId, (err) => {
+            if (!err) {
+                console.log("Successfully Deleted Item!")
+                res.redirect('/')
+            }
+        })
+
+    } else {
+        List.findOneAndUpdate({name: listName},{$pull:{items:{_id: checkedItemId}}},(err,foundList)=>{
+            if (!err) {
+                res.redirect("/"+listName);
+            }
+        })
+        
+    }
+    
 })
 
 app.listen(PORT, () => console.log("Server running on port " + PORT));
